Task: Update Dependencies
Updates project dependencies to the latest versions. This update includes bug fixes, performance improvements, and new features for the libraries used.

Updated dependencies include, but are not limited to:
- axe-core
- function-binding
- generator-function
- is-regex
- is-symbol
- inspect-object
- typescript
- zod
